Massive rewrite of nvim plugins (Bad commit etiquette)
