--- nvim/lua/haskell.lua
@@ -1,4 +1,0 @@
-local lspconfig = require "lspconfig"
-lspconfig.hls.setup({
-    cmd = { "ghcup run hls", "--lsp" },
-})
