+++ nvim/lua/plugins/conform.lua
@@ -32,12 +32,5 @@ return {
                 timeout_ms = 500,
             },
         })
-        vim.keymap.set({ "n", "v" }, "<leader>mp", function()
-            conform.format({
-                lsp_fallback = true,
-                async = false,
-                timeout_ms = 500,
-            })
-        end, { desc = "Format file or range (in visual mode)" })
     end,
 }
