+++ nvim/lua/plugins/context.lua
@@ -1,0 +1,3 @@
+return {
+    "nvim-treesitter/nvim-treesitter-context",
+}
