+++ nvim/lua/plugins/dap.lua
@@ -1,5 +1,4 @@
 return {
-
     {
         "jay-babu/mason-nvim-dap.nvim",
         event = "VeryLazy",
@@ -19,16 +18,18 @@ return {
         },
         event = "VeryLazy",
         config = function()
-            local dap = require "dap"
-            local dapui = require "dapui"
+            local dap, dapui = require "dap", require "dapui"
             dapui.setup()
-            dap.listeners.after.event_initialized["dapui_config"] = function()
+            dap.listeners.before["attach"]["dapui_config"] = function()
                 dapui.open()
             end
-            dap.listeners.after.event_terminated["dapui_config"] = function()
+            dap.listeners.before["launch"]["dapui_config"] = function()
+                dapui.open()
+            end
+            dap.listeners.before["event_terminated"]["dapui_config"] = function()
                 dapui.close()
             end
-            dap.listeners.after.event_exited["dapui_config"] = function()
+            dap.listeners.before["event_exited"]["dapui_config"] = function()
                 dapui.close()
             end
         end,
