+++ nvim/lua/plugins/markview.lua
@@ -1,0 +1,12 @@
+return {
+    "OXY2DEV/markview.nvim",
+    ft = "markdown",
+    dependencies = {
+        -- You will not need this if you installed the
+        -- parsers manually
+        -- Or if the parsers are in your $RUNTIMEPATH
+        "nvim-treesitter/nvim-treesitter",
+
+        "nvim-tree/nvim-web-devicons",
+    },
+}
