+++ nvim/lua/plugins/neorg.lua
@@ -40,7 +40,7 @@ return {
                 },
             },
         })
-        vim.wo.foldlevel = 99
-        vim.wo.conceallevel = 2
     end,
+    ft = "norg",
+    cmd = "Neorg",
 }
