+++ nvim/lua/plugins/nvim-lint.lua
@@ -9,11 +9,6 @@ return {
         local cpplint = lint.linters.cpplint
         cpplint.args = { "--filter=-build/c++11,-build/c++14,-build/c++tr1" }
         lint.linters_by_ft = {
-            javascript = { "eslint_d" },
-            typescript = { "eslint_d" },
-            javascriptreact = { "eslint_d" },
-            typescriptreact = { "eslint_d" },
-            svelte = { "eslint_d" },
             python = { "mypy", "ruff" },
             cpp = { "cpplint" },
         }
