+++ nvim/lua/plugins/overseer.lua
@@ -1,0 +1,4 @@
+return {
+    "stevearc/overseer.nvim",
+    opts = {},
+}
