+++ nvim/lua/plugins/telescope.lua
@@ -1,0 +1,60 @@
+return {
+    {
+        "nvim-telescope/telescope.nvim",
+        branch = "0.1.x",
+        dependencies = {
+            "nvim-lua/plenary.nvim",
+            -- Fuzzy Finder Algorithm which requires local dependencies to be built.
+            -- Only load if `make` is available. Make sure you have the system
+            -- requirements installed.
+            {
+                "nvim-telescope/telescope-fzf-native.nvim",
+                -- NOTE: If you are having trouble with this installation,
+                --       refer to the README for telescope-fzf-native for more instructions.
+                build = "make",
+                cond = function()
+                    return vim.fn.executable "make" == 1
+                end,
+            },
+            { "nvim-telescope/telescope-ui-select.nvim" },
+
+            "debugloop/telescope-undo.nvim",
+
+            -- Useful for getting pretty icons, but requires a Nerd Font.
+            { "nvim-tree/nvim-web-devicons", enabled = vim.g.have_nerd_font },
+        },
+        config = function()
+            -- [[ Configure Telescope ]]
+            -- See `:help telescope` and `:help telescope.setup()`
+            require("telescope").setup({
+                -- You can put your default mappings / updates / etc. in here
+                --  All the info you're looking for is in `:help telescope.setup()`
+                --
+                defaults = {
+                    mappings = {
+                        i = {
+                            ["<C-u>"] = false,
+                            ["<C-d>"] = false,
+                        },
+                    },
+                },
+                -- defaults = {
+                --   mappings = {
+                --     i = { ['<c-enter>'] = 'to_fuzzy_refine' },
+                --   },
+                -- },
+                -- pickers = {}
+                extensions = {
+                    ["ui-select"] = {
+                        require("telescope.themes").get_dropdown(),
+                    },
+                    undo = {},
+                },
+            })
+            -- Enable telescope fzf native, if installed
+            pcall(require("telescope").load_extension, "fzf")
+            pcall(require("telescope").load_extension, "ui-select")
+            pcall(require("telescope").load_extension, "undo")
+        end,
+    },
+}
