+++ nvim/lua/plugins/vimbegood.lua
@@ -1,3 +1,5 @@
 return {
     "ThePrimeagen/vim-be-good",
+    cmd = "VimBeGood",
+    event = "VeryLazy",
 }
